+++ src/main/java/inheritance/Car.java
@@ -1,3 +1,5 @@
+package inheritance;
+
 public abstract class Car {
     /**
      * 리터당 이동 거리. 즉, 연비
