+++ src/test/java/inheritance/RentCompanyTest.java
@@ -1,3 +1,5 @@
+package inheritance;
+
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -16,11 +18,11 @@ class RentCompanyTest {
 
         String report = company.generateReport();
         assertThat(report).isEqualTo(
-                "Sonata : 15리터" + NEWLINE +
-                "K5 : 20리터" + NEWLINE +
-                "Sonata : 12리터" + NEWLINE +
-                "Avante : 20리터" + NEWLINE +
-                "K5 : 30리터" + NEWLINE
+                "inheritance.Sonata : 15리터" + NEWLINE +
+                "inheritance.K5 : 20리터" + NEWLINE +
+                "inheritance.Sonata : 12리터" + NEWLINE +
+                "inheritance.Avante : 20리터" + NEWLINE +
+                "inheritance.K5 : 30리터" + NEWLINE
         );
     }
 }
